pyproject.toml
[project]
name = "google-search-api-streamit"
version = "0.1.0"
description = "간단한 Gemini File Search 기반 문서 질의응답 챗봇"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.28.0",
    "google-genai>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
target-version = "py39"
